Use python-gammu from PyPI
The python3-gammu package is missing from Debian Bullseye, and so are
several of the underlying gammu library packages. Most of them are
available again in the bullseye-backports archive, but not the Python
libraries. Everything is available in Sid.

These changes ensure we have the necessary libraries from
bullseye-backports to build and install python-gammu directly in NAV's
virtualenv.