Enable include-system-site-packages the proper way
On Python 3.7 and earlier, inclusion of global site packages into
virtualenvs was achieved by deleting a flag file.  On Python 3.9, this
is now accomplished by changing a setting in the virtual environment's
`pyvenv.cfg` configuration file.

Some system package imports would not work for NAV integrations on
Debian Bullseye until this was fixed.