Fix broken addition of navcron into dialout
navcron must be added to the dialout group in order for gammu SMS
dispatch on locally connected GSM devices to work (the dialout group
has write privileges to serial port devices on the system).

However, the correct way to do this today is with the `adduser` command,
not the `addgroup` command, which fails under bookworm when used like
this would before.